Converting to MVC... sort of?
I'm... still not 100% sure how Model View Control is supposed to work, but from what I understand, this should be it...?
